requirements.txt
# pip install --requirement=requirements.txt
django==1.3.1
coverage==3.5.1
django_nose==0.1.3
